Implement RaceCountdown, fix bugs, enable input
Fix some memory not being initialized, such as the bool that disbles
input. So the input was previously enabled, but by error.
In this patch we properly disable the input and implement the
RaceCountdown class in charge of enabling it.

The game is now "playable" with a very very broad definition of fun.